Move Redis related code to `redis`
